* 'develop' of https://gitlab.com/jusfoun-bigdata/hookah:
  发布商品BUG
  属性  类型更改
  fix bug
  字段调整
  出入金

+++ hookah-pay/src/main/java/com.jusfoun.hookah.pay/controller/TestController.java
@@ -27,7 +27,7 @@ public class TestController {
         MoneyInOutBo moneyInOutBo = new MoneyInOutBo();
         moneyInOutBo.setMoney(1000L);
         moneyInOutBo.setPayAccountID(2l);
-        moneyInOutBo.setOperatorType(1);
+        moneyInOutBo.setOperatorType((byte)1);
         moneyInOutBo.setUserId("1234567");
         payAccountRecordService.entryAndExitPayments(moneyInOutBo);
     }
